Port TestYearMonth to scalatest
